14001: camera_v2: Forward-port from 3.4 kernel
*Currently causes SMD crash immediately when opening the camera

+++ drivers/media/platform/msm/camera_v2/sensor/csiphy/msm_csiphy.h
@@ -42,10 +42,8 @@ struct csiphy_device {
 	struct mutex mutex;
 	uint32_t hw_version;
 	enum msm_csiphy_state_t csiphy_state;
-
 	struct clk *csiphy_clk[CSIPHY_NUM_CLK_MAX];
-
-	int32_t ref_count;
+	uint8_t ref_count;
 	uint16_t lane_mask[MAX_CSIPHY];
 };
